refacto: proximity analysis' initial filter
