+++ app/src/main/java/com/illiouchine/jm/ui/composable/plot/ProximitySpider.kt
@@ -19,15 +19,14 @@ import androidx.compose.ui.text.TextStyle
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.illiouchine.jm.extensions.shortenNames
 import com.illiouchine.jm.service.ProximityAnalysis
 import com.illiouchine.jm.service.ProximityAnalyzer
 import com.illiouchine.jm.ui.composable.plot.lib.SpiderChart
-import com.illiouchine.jm.ui.composable.plot.utils.truncate
+import com.illiouchine.jm.ui.composable.plot.utils.filterAnalysis
+import com.illiouchine.jm.ui.composable.plot.utils.makeProposalsInitials
 import com.illiouchine.jm.ui.preview.PreviewDataFaker
 import com.illiouchine.jm.ui.theme.JmTheme
 import com.illiouchine.jm.ui.theme.Theme
-import java.lang.Integer.min
 
 // Shows how close (in the collective hearts of the judges) pairs of proposals are.
 // A proximity of +1 means that the two proposals received exactly the same grades in each ballot.
@@ -44,34 +43,20 @@ fun ProximitySpider(
     // FIXME: move this out of here (use params)
     var selectedCategoryIndex by remember { mutableIntStateOf(0) }
 
-    // FIXME: duplicated code with the bar chart
-    val allProposalsIndices = 0.rangeUntil(analysis.proposals.size).toList()
-    val lotsOfProposalsIndices = proposalsIndices ?: allProposalsIndices
-
-    val maxAmountOfProposalsThatFit = 16
-    val maxAmountOfProposals = min(maxAmountOfProposalsThatFit, lotsOfProposalsIndices.size)
-
-    val usedProposalsIndices = lotsOfProposalsIndices.take(maxAmountOfProposals)
-    val filteredAnalysis = analysis.filterByProposalsIndices(usedProposalsIndices)
-
-    val proposalsInitials = filteredAnalysis.proposals.shortenNames().map {
-        it.truncate(
-            maxLength = 7, // check big fonts on small screens if you increment this
-            ellipsis = "…",
-        )
-    }
+    val usedAnalysis = filterAnalysis(analysis, proposalsIndices)
+    val proposalsInitials = makeProposalsInitials(usedAnalysis)
 
     SpiderChart(
         modifier = modifier,
         title = {
-            Text("Proximity with ${filteredAnalysis.proposals[selectedCategoryIndex]}")
+            Text("Proximity with ${usedAnalysis.proposals[selectedCategoryIndex]}")
         },
         // We are not using the legend because it yields a (min > max) error from the Koala lib.
         // legend = {},
         // Truncating is not enough on small screens with big fonts — responsive for tablets ?
         // categories = filteredAnalysis.proposals.map { it.truncate(16, "…") },
         categories = proposalsInitials, // so we use initials, it worked nicely so far
-        values = filteredAnalysis.proximities[selectedCategoryIndex].map { it.toFloat() },
+        values = usedAnalysis.proximities[selectedCategoryIndex].map { it.toFloat() },
         tickValues = listOf(-1f, 0f, 1f),
         tickDecimals = 0,
         highlightedCategoryIndex = selectedCategoryIndex,
@@ -86,7 +71,7 @@ fun ProximitySpider(
             modifier = Modifier.align(alignment = Alignment.CenterHorizontally)
         ) {
             Column {
-                filteredAnalysis.proposals.forEachIndexed { index, name ->
+                usedAnalysis.proposals.forEachIndexed { index, name ->
                     Row {
                         Text(
                             modifier = Modifier
